Added the food aisle into unreal
I added a new scene along with its own textures and assets. Needs a litte work, but is overall in a good state. I used the toy scene floor, and need a cieling that the other grocery store scenes use. Textures will be reworked, and food items will be implemented.